+++ src/App1.js
@@ -1,23 +1,16 @@
-This Example shows the creation of a simple __Viewer Widget__.
-
-```js
 import {useCallback, useState, useMemo} from 'react';
-import ViewerWidget from './ViewerWidget';
-import AboutButton from '../../../button/common/AboutButton/AboutButton';
-import {Button, Layout} from 'antd';
+import ViewerWidget from './components/widget/viewer/ViewerWidget/ViewerWidget'
+import AboutButton from './components/button/common/AboutButton/AboutButton';	
 import { Ion, defined, Cartesian3, Math, GoogleMaps, createGooglePhotorealistic3DTileset } from 'cesium';
-import 'cesium/Widgets/widgets.css'
-import '../../../../assets/css/react-cesiumext-controls.css';
-const { Content } = Layout;
+import './assets/css/react-cesiumext-controls.css';
 
-//register and get your own key for CesiumJs Ion
 Ion.defaultAccessToken = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJqdGkiOiI1OGZjNDZkNC1iOTdlLTRhYWMtODBjYy1mNWIzOGEwYjUxNjAiLCJpZCI6MTAzODcsInNjb3BlcyI6WyJhc3IiLCJnYyJdLCJpYXQiOjE1NTYyODk0MDl9.f13hGNgcrSFUzcocb5CpHD3Im9xzT0c7IDAPcpwGidc';
 
-//this api key is only valid for this application. Create your own googleMaps api key
-GoogleMaps.defaultApiKey = 'AIzaSyBzf8jvuiGJR0x7q2eqaILi4FzKwAq8xXA';
+GoogleMaps.defaultApiKey = 'your google api key';
 
-const ViewerWidgetExample = () => {
+const App1 = () => {
    
+    console.log('GoogleMaps.mapTilesApiEndpoint', GoogleMaps.mapTilesApiEndpoint);
     const [viewer, setViewer] = useState(null);
 
     const viewerOpts = useMemo(() => {
@@ -30,20 +23,22 @@ const ViewerWidgetExample = () => {
     const onStart = useCallback((aViewer) => {
         console.log(aViewer);
         aViewer.scene.globe.show = false;
-        //Add Google Photorealistic 3D Tileset
+        // Add Google Photorealistic 3D Tileset
         createGooglePhotorealistic3DTileset()
         .then((tileset) => {
             if(defined(aViewer) && !aViewer.isDestroyed()) {
                 aViewer.scene.primitives.add(tileset);
                 console.log('tileset added');
             }
+            else {
+                console.log('tileset not added');
+            }
         })
         .catch((error) => {
             console.log(`Error creating tileset: ${error}`);
         }); 
 
-        // Flying the camera near to Proximus Tower, Belgium ... 
-        // and maybe u can find me there :-)
+        // Fly the camera to San Francisco at the given longitude, latitude, and height.
         aViewer.camera.flyTo({
             destination : Cartesian3.fromDegrees(4.358948, 50.859114, 200),
             orientation : {
@@ -55,15 +50,9 @@ const ViewerWidgetExample = () => {
     }, []);
     
     return (
-        <Layout>
-            <Content>
-                <ViewerWidget options={viewerOpts} onStart={onStart}>
-                    <AboutButton className="react-cesiumext-control" style={{top:4, left:4}}/>
-                </ViewerWidget>
-            </Content>
-        </Layout>
+      <ViewerWidget options={viewerOpts} onStart={onStart}>
+        <AboutButton className="react-cesiumext-control" style={{top:4, left:4}}/>
+      </ViewerWidget>
     );
-}
-
-<ViewerWidgetExample/>
-```
+};
+export default App1;