+++ src/components/grid/dropdown/EntityGridOptionsDropdown/EntityGridOptionsDropdown.js
@@ -2,12 +2,14 @@ import { useMemo } from 'react';
 import { Button, Dropdown} from 'antd';
 import UnloadDataSourceButton from '../../../button/dataSource/UnloadDataSourceButton/UnloadDataSourceButton';
 import FlyToButton from '../../../button/viewer/FlyToButton/FlyToButton';
+import ExportCsvFromGridButton from '../../button/ExportCsvFromGridButton/ExportCsvFromGridButton';
 import { defined } from 'cesium';
 
 const EntityGridOptionsDropdown = ({
     viewer,
     ds,
     selectedEntities,
+    gridApi,
     children="Options",
     menuPropsFunc,
     ...otherProps
@@ -65,6 +67,25 @@ const EntityGridOptionsDropdown = ({
             }
         );
         menuProps.items.push(zoomItems);
+        //Export options
+        const exportItems = {
+            key: 'EXPORT',
+            label: 'Export',
+            children:[]
+        };
+        exportItems.children.push(
+            {
+                key: 'EXPORT_CSV',
+                label:  <ExportCsvFromGridButton 
+                            size="small" 
+                            type="text" 
+                            gridApi={gridApi}
+                        >
+                            Export CSV
+                        </ExportCsvFromGridButton>
+            },
+        );
+        menuProps.items.push(exportItems);
 
         if(defined(menuPropsFunc)) {
             const additionalItems = menuPropsFunc(viewer, ds, selectedEntities);
@@ -72,7 +93,7 @@ const EntityGridOptionsDropdown = ({
         }
 
         return menuProps;
-    }, [ds, viewer, menuPropsFunc, selectedEntities]);
+    }, [ds, viewer, menuPropsFunc, selectedEntities, gridApi]);
 
     return (
         <Dropdown menu = {menu} placement="bottomLeft" trigger="click">