+++ CHANGELOG.md
@@ -3,8 +3,15 @@
 ## v0.1.2 - 05/04/2023
 - First Beta release
 
-## v0.1.3 - xx/xx/2023
+## v0.1.3 - 04/05/2023
 - Upgraded react-styleguidist from version 11.2.0 to version 13.1.1
 - Upgraded ag-grid-react and ag-grid-community from version 28.x to version 29.3.0
 - Removed the call for all the depracated Cesium functions.
 - cleared the javascript 'console.log' from code
+
+## v0.1.4 - 15/05/2023
+- Upgraded ag-grid-react and ag-grid-community from 29.3.0 to version 29.3.4
+- Upgraded antd to version 5.4.7
+- Upgraded Cesium to 1.105.2 with support to GoogleMaps Photorealistic 3D Tiles
+- Cesium added as a per-dependency
+- Entity Grid has has the option to export the data to CSV file